# perf(rolldown): use starts_with instead of regex in resolve